+++ aim/cli/version/commands.py
@@ -5,4 +5,4 @@ from aim.__version__ import __version__ as aim_version
 
 @click.command()
 def version():
-    click.echo('aim-cli v{}'.format(aim_version))
+    click.echo('Aim v{}'.format(aim_version))